requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.6
    # via langchain
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpx
    #   langchain
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
cryptography==41.0.5
    # via pyjwt
dataclasses-json==0.6.2
    # via langchain
deprecated==1.2.14
    # via pygithub
distro==1.8.0
    # via openai
exceptiongroup==1.1.3
    # via anyio
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.1
    # via openai
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.334
    # via -r requirements.in
langsmith==0.0.63
    # via langchain
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.1
    # via langchain
openai==1.2.3
    # via -r requirements.in
packaging==23.2
    # via marshmallow
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via
    #   langchain
    #   langsmith
    #   openai
pydantic-core==2.10.1
    # via pydantic
pygithub==2.1.1
    # via -r requirements.in
pyjwt[crypto]==2.8.0
    # via
    #   pygithub
    #   pyjwt
pynacl==1.5.0
    # via pygithub
python-dateutil==2.8.2
    # via pygithub
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via langchain
regex==2023.10.3
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langsmith
    #   pygithub
    #   tiktoken
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.23
    # via langchain
tenacity==8.2.3
    # via langchain
tiktoken==0.5.1
    # via -r requirements.in
tqdm==4.66.1
    # via openai
typing-extensions==4.8.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.0.7
    # via
    #   pygithub
    #   requests
wrapt==1.16.0
    # via deprecated
yarl==1.9.2
    # via aiohttp
